refactor: extract auth link helper to shared module
Move createAuthLink function from auth-recovery.spec.ts to
helpers.ts as getCapturedAuthLink to reduce code duplication
across test files. Update auth.spec.ts to use the shared helper
and expand unverified email test to verify recovery flow works.
