+++ e2e/auth-recovery.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test, type APIRequestContext } from "@playwright/test";
+import { expect, test } from "@playwright/test";
 import {
   E2E_CHANGE_EMAIL_DISPLAY_NAME,
   E2E_CHANGE_EMAIL_EMAIL,
@@ -9,33 +9,7 @@ import {
   E2E_REGISTER_FLOW_NEW_PASSWORD,
   E2E_REGISTER_FLOW_PASSWORD,
 } from "./global-setup";
-import { signIn, signOut } from "./helpers";
-
-async function createAuthLink(
-  request: APIRequestContext,
-  kind: "verification" | "change_email_verification" | "password_reset",
-  email: string,
-): Promise<string> {
-  const deadline = Date.now() + 15000;
-
-  while (Date.now() < deadline) {
-    const response = await request.post("/api/v1/test/auth-links", {
-      data: { kind, email },
-    });
-
-    if (response.ok()) {
-      const body = (await response.json()) as { data?: { url?: string } };
-      const url = body.data?.url;
-
-      expect(url).toBeTruthy();
-      return url as string;
-    }
-
-    await new Promise((resolvePromise) => setTimeout(resolvePromise, 250));
-  }
-
-  throw new Error(`Timed out creating ${kind} auth link for ${email}.`);
-}
+import { getCapturedAuthLink, signIn, signOut } from "./helpers";
 
 test("user can register, verify email, request a password reset, and sign in with the new password", async ({
   page,
@@ -47,7 +21,11 @@ test("user can register, verify email, request a password reset, and sign in wit
   await page.fill("#register-password", E2E_REGISTER_FLOW_PASSWORD);
   await page.getByRole("button", { name: "Create Account" }).click();
 
-  const verificationUrl = await createAuthLink(request, "verification", E2E_REGISTER_FLOW_EMAIL);
+  const verificationUrl = await getCapturedAuthLink(
+    request,
+    "verification",
+    E2E_REGISTER_FLOW_EMAIL,
+  );
   await page.goto(verificationUrl);
   await page.waitForURL("/");
 
@@ -60,7 +38,7 @@ test("user can register, verify email, request a password reset, and sign in wit
   await page.fill("#forgot-email", E2E_REGISTER_FLOW_EMAIL);
   await page.getByRole("button", { name: "Send Reset Link" }).click();
 
-  const resetUrl = await createAuthLink(request, "password_reset", E2E_REGISTER_FLOW_EMAIL);
+  const resetUrl = await getCapturedAuthLink(request, "password_reset", E2E_REGISTER_FLOW_EMAIL);
   await page.goto(resetUrl);
   await expect(page.getByRole("heading", { name: "Set New Password" })).toBeVisible();
 
@@ -85,7 +63,7 @@ test("user can request and verify an email change", async ({ page, request }) =>
 
   await expect(page.getByRole("status")).toContainText(E2E_CHANGE_EMAIL_NEW_EMAIL);
 
-  const verificationUrl = await createAuthLink(
+  const verificationUrl = await getCapturedAuthLink(
     request,
     "change_email_verification",
     E2E_CHANGE_EMAIL_NEW_EMAIL,
